// src/interfaces/Candidate.interface.tsx
export interface Candidate {
    // Identifiers
    login: string;
    id: number;
    node_id: string;
    avatar_url: string;
    gravatar_id: string;
    type: string;
    site_admin: boolean;

    // Profile
    name?: string | null;
    company?: string | null;
    blog?: string | null;
    location?: string | null;
    email?: string | null;
    hireable?: boolean | null;
    bio?: string | null;
    twitter_username?: string | null;

    // Statistics
    public_repos: number;
    public_gists: number;
    followers: number;
    following: number;

    // Dates
    created_at: string;
    updated_at: string;

    // URLs
    url: string;
    html_url: string;
    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscriptions_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_events_url: string;
}